// src/app/[page]/components/categories/types.ts
export type MobileCardVariant =
  | "text"
  | "chip"
  | "rectChip"
  | "imageTop"
  | "circleImageTop"
  | "imageOverlay"
  | "image4Over";

export type MobileSizeVariant = "mini" | "sm" | "md" | "lg" | "xl" | "big";
export type MobileLayoutVariant = "default" | "full" | "scroll";
export type DesktopCategoriesVariant =
  | "default"
  | "pill"
  | "notpill"
  | "bigcategory"
  | "compact";